tests/wpt/tests/encoding-detection/zh-Big5-late.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>zh Big5 late</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "Big5", 'Expected Big5');
    }, "Check detection result");
    w.close();
    done();
};
var w = window.open("support/zh-Big5-late.sub.html");
</script>
